Revert "style: formatting"
This reverts commit cbc8ab27bdb55acef9a760836389d500f9cb5f62.

+++ src/utils/functions/moderation/ban.ts
@@ -11,14 +11,8 @@ export async function newBan(guild: Guild, userIDs: string[] | string, date: Dat
   }
 
   for (const userID of userIDs) {
-    if (
-      await prisma.moderationBan.findUnique({
-        where: { userId_guildId: { guildId: guild.id, userId: userID } },
-      })
-    )
-      await prisma.moderationBan.delete({
-        where: { userId_guildId: { guildId: guild.id, userId: userID } },
-      });
+    if (await prisma.moderationBan.findUnique({ where: { userId_guildId: { guildId: guild.id, userId: userID } } }))
+      await prisma.moderationBan.delete({ where: { userId_guildId: { guildId: guild.id, userId: userID } } });
 
     await prisma.moderationBan.create({
       data: {
